Move the demoMode file
